requirements.txt
discord.py[voice]>=2.3.0
yt-dlp>=2023.7.6
PyNaCl>=1.5.0
python-dotenv>=1.0.0
spotipy>=2.22.1
requests>=2.31.0
